+++ src/main/resources/clcode/default/headers/matrixmath.h
@@ -8,21 +8,21 @@
     r[2].x r[2].y r[2].y
 */
 struct matrix3x3 {
-    double3 r[3];
+    float3 r[3];
 };
 
 struct matrix3x3 matrixProduct(struct matrix3x3 a, struct matrix3x3 b);
 
-double3 matrixTimesVector(struct matrix3x3 m, double3 vector);
+float3 matrixTimesVector(struct matrix3x3 m, float3 vector);
 
-struct matrix3x3 rotationMatrix(double alpha, double beta, double gamma);
+struct matrix3x3 rotationMatrix(float alpha, float beta, float gamma);
 
-struct matrix3x3 reverseRotationMatrix(double gamma, double beta, double alpha);
+struct matrix3x3 reverseRotationMatrix(float gamma, float beta, float alpha);
 
-struct matrix3x3 rotationMatrixX(double alpha);
+struct matrix3x3 rotationMatrixX(float alpha);
 
-struct matrix3x3 rotationMatrixY(double beta);
+struct matrix3x3 rotationMatrixY(float beta);
 
-struct matrix3x3 rotationMatrixZ(double gamma);
+struct matrix3x3 rotationMatrixZ(float gamma);
 
 #endif
